+++ src/promql/src/functions/resets.rs
@@ -37,63 +37,28 @@ impl Resets {
 mod test {
     use std::sync::Arc;
 
-    use datafusion::arrow::array::{Array, Float64Array, TimestampMillisecondArray};
+    use datafusion::arrow::array::{Float64Array, TimestampMillisecondArray};
     use datafusion::arrow::buffer::NullBuffer;
-    use datafusion::common::DataFusionError;
-    use datafusion::physical_plan::ColumnarValue;
-    use datafusion_common::config::ConfigOptions;
-    use datafusion_expr::ScalarFunctionArgs;
-    use datatypes::arrow::datatypes::{DataType, Field};
+    use datatypes::arrow::array::Array;
 
     use super::*;
-    use crate::functions::test_util::simple_range_udf_runner;
+    use crate::functions::test_util::{
+        self, STALE_NAN, TinyPrng, assert_execution_error, build_test_range_arrays,
+        invoke_range_udf, simple_range_udf_runner,
+    };
     use crate::range_array::RangeArray;
 
-    // build timestamp range and value range arrays for test
-    fn build_test_range_arrays(
-        timestamps: Vec<i64>,
-        values: Vec<f64>,
-        ranges: Vec<(u32, u32)>,
-    ) -> (RangeArray, RangeArray) {
-        let ts_array = Arc::new(TimestampMillisecondArray::from_iter(
-            timestamps.into_iter().map(Some),
-        ));
-        let values_array = Arc::new(Float64Array::from_iter(values));
-
-        let ts_range_array = RangeArray::from_ranges(ts_array, ranges.clone()).unwrap();
-        let value_range_array = RangeArray::from_ranges(values_array, ranges).unwrap();
-
-        (ts_range_array, value_range_array)
-    }
-
-    fn invoke_range_udf(
-        udf: ScalarUDF,
-        timestamps: RangeArray,
-        values: RangeArray,
-    ) -> Result<ColumnarValue, DataFusionError> {
-        let number_rows = timestamps.len();
-        let args = vec![
-            ColumnarValue::Array(Arc::new(timestamps.into_dict())),
-            ColumnarValue::Array(Arc::new(values.into_dict())),
-        ];
-        let arg_fields = vec![
-            Arc::new(Field::new("timestamps", args[0].data_type(), false)),
-            Arc::new(Field::new("values", args[1].data_type(), false)),
-        ];
-        udf.invoke_with_args(ScalarFunctionArgs {
-            args,
-            arg_fields,
-            number_rows,
-            return_field: Arc::new(Field::new("result", DataType::Float64, false)),
-            config_options: Arc::new(ConfigOptions::default()),
-        })
-    }
-
-    fn assert_execution_error(error: DataFusionError, expected: &str) {
-        match error {
-            DataFusionError::Execution(message) => assert_eq!(message, expected),
-            other => panic!("expected execution error, got {other:?}"),
+    fn resets_oracle(values: &[f64]) -> Option<f64> {
+        let (first, rest) = values.split_first()?;
+        let mut resets = 0;
+        let mut previous = first;
+        for current in rest {
+            if current < previous {
+                resets += 1;
+            }
+            previous = current;
         }
+        Some(resets as f64)
     }
 
     #[test]
@@ -145,87 +110,6 @@ mod test {
         );
     }
 
-    const STALE_NAN: f64 = f64::from_bits(0x7ff0_0000_0000_0002);
-
-    struct TinyPrng(u64);
-
-    impl TinyPrng {
-        fn next_u32(&mut self) -> u32 {
-            self.0 ^= self.0 << 13;
-            self.0 ^= self.0 >> 7;
-            self.0 ^= self.0 << 17;
-            self.0 as u32
-        }
-
-        fn next_index(&mut self, bound: usize) -> usize {
-            (self.next_u32() as usize) % bound
-        }
-    }
-
-    fn resets_oracle(values: &[f64]) -> Option<f64> {
-        let (first, rest) = values.split_first()?;
-        let mut resets = 0;
-        let mut previous = first;
-        for current in rest {
-            if current < previous {
-                resets += 1;
-            }
-            previous = current;
-        }
-        Some(resets as f64)
-    }
-
-    fn run_oracle_ranges(
-        values: Vec<Option<f64>>,
-        raw_values: Vec<f64>,
-        timestamp_ranges: Vec<(u32, u32)>,
-        value_ranges: Vec<(u32, u32)>,
-    ) -> Vec<Option<f64>> {
-        assert_eq!(timestamp_ranges.len(), value_ranges.len());
-        assert!(
-            timestamp_ranges
-                .iter()
-                .zip(&value_ranges)
-                .all(|((_, timestamp_length), (_, value_length))| timestamp_length == value_length)
-        );
-        assert_eq!(values.len(), raw_values.len());
-        let nulls = values.iter().map(Option::is_none).collect::<Vec<_>>();
-
-        let expected = value_ranges
-            .iter()
-            .map(|(offset, length)| {
-                resets_oracle(&raw_values[*offset as usize..(*offset + *length) as usize])
-            })
-            .collect::<Vec<_>>();
-
-        let timestamp_values = (0..64)
-            .map(|value| Some(i64::from(value) * 1_000))
-            .collect::<Vec<_>>();
-        let timestamp_array = Arc::new(TimestampMillisecondArray::from_iter(timestamp_values));
-        let value_array = Arc::new(Float64Array::from_iter(values));
-        for (index, ((actual, expected), is_null)) in value_array
-            .values()
-            .iter()
-            .zip(&raw_values)
-            .zip(nulls)
-            .enumerate()
-        {
-            assert_eq!(actual.to_bits(), expected.to_bits());
-            assert_eq!(value_array.is_null(index), is_null);
-        }
-
-        let timestamp_ranges = RangeArray::from_ranges(timestamp_array, timestamp_ranges).unwrap();
-        let value_ranges = RangeArray::from_ranges(value_array, value_ranges).unwrap();
-        simple_range_udf_runner(
-            Resets::scalar_udf(),
-            timestamp_ranges,
-            value_ranges,
-            vec![],
-            expected.clone(),
-        );
-        expected
-    }
-
     #[test]
     fn resets_range_array_oracle_edge_cases() {
         let values = vec![
@@ -246,22 +130,26 @@ mod test {
             Some(0.0),
         ];
         let raw_values = values.iter().map(|value| value.unwrap()).collect();
-        let expected = run_oracle_ranges(
+        let expected = test_util::run_oracle_ranges(
             values,
             raw_values,
             vec![(5, 0), (2, 1), (7, 5), (0, 5), (11, 5), (16, 4)],
             vec![(0, 0), (0, 1), (4, 5), (0, 5), (8, 5), (1, 4)],
+            resets_oracle,
+            Resets::scalar_udf(),
         );
         assert_eq!(
             expected,
             vec![None, Some(0.0), Some(0.0), Some(1.0), Some(2.0), Some(1.0)]
         );
 
-        let expected = run_oracle_ranges(
+        let expected = test_util::run_oracle_ranges(
             vec![Some(2.0), None, Some(2.0)],
             vec![2.0, 0.0, 2.0],
             vec![(9, 3), (1, 1)],
             vec![(0, 3), (1, 1)],
+            resets_oracle,
+            Resets::scalar_udf(),
         );
         assert_eq!(expected, vec![Some(1.0), Some(0.0)]);
     }
@@ -292,7 +180,14 @@ mod test {
             value_ranges.push((prng.next_index(49 - length as usize) as u32, length));
         }
 
-        run_oracle_ranges(values, raw_values, timestamp_ranges, value_ranges);
+        test_util::run_oracle_ranges(
+            values,
+            raw_values,
+            timestamp_ranges,
+            value_ranges,
+            resets_oracle,
+            Resets::scalar_udf(),
+        );
     }
 
     #[test]
